wol-service/requirements.txt
flask
pywakeonlan
regex